Create JobStore service
